Add a first file .
It contains code from scratch of a perceptron. In this code we created a dataset from the make_classification class imported from the sklearn library we take all rows one by one and find the actual weights and bias values that divide that data. Finally, we find the actual line and then we plot it using the matplotlib library.